feat+perf: work-in-progress - more marshalling impl. including unmanaged-to-managed - update issue list - expose a method for allocation-free construction from string literals that has better UX - optimize cast to ReadOnlyMemory<byte> - move literal pool to a separate type to reuse it - implement specialized runes/char join paths which may be a somewhat common operation to do that would be quite slow if it goes through boxing and virtual calls when being passed to high level IEnumerable overload - now very fast although reach bikeshedding level previously considered impossible - marginally improve codegen for VectorExtensions - documentation

+++ Sources/Comparers/U8AsciiIgnoreCaseComparer.cs
@@ -64,7 +64,7 @@ public readonly struct U8AsciiIgnoreCaseComparer : IU8Comparer
                 var neqmask = ~((lvec | lcmask).Eq(rvec | rcmask));
                 if (neqmask != Vector256<byte>.Zero)
                 {
-                    return neqmask.IndexOfMatch() + (int)offset;
+                    return (int)(uint)(neqmask.IndexOfMatch() + offset);
                 }
                 offset += (nuint)Vector256<byte>.Count;
             } while (offset <= lastvec);
@@ -81,7 +81,7 @@ public readonly struct U8AsciiIgnoreCaseComparer : IU8Comparer
             var neqmask = ~lcvec.Eq(rcvec);
             if (neqmask != Vector128<byte>.Zero)
             {
-                return neqmask.IndexOfMatch() + (int)offset;
+                return (int)(uint)(neqmask.IndexOfMatch() + offset);
             }
             offset += (nuint)Vector128<byte>.Count;
         }
